# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Catalog Service API
  description: |
    Manages the RUBY MUSIC catalog: genres, stations, artists, albums and songs.
    Stations and Genres are distinct concepts: a Genre is a musical tag on songs;
    a Station is a themed listening room that aggregates songs from one or more genres.
    All write operations require admin privileges (enforced at gateway level).
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local development
  - url: http://localhost:8080/api/v1/catalog
    description: Via API Gateway
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/catalog/genres:
    get:
      tags: [Genres]
      operationId: listGenres
      summary: List all genres (optionally filtered by name — admin search bar)
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Filter by name (partial match, case-insensitive)
      responses:
        '200':
          description: List of genres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GenreResponse'
    post:
      tags: [Genres]
      operationId: createGenre
      summary: Create a genre (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreRequest'
      responses:
        '201':
          description: Genre created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreResponse'
        '409':
          description: Genre name already exists
  /api/v1/catalog/genres/{id}:
    get:
      tags: [Genres]
      operationId: getGenreById
      summary: Get genre by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Genre found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreResponse'
        '404':
          description: Not found
    put:
      tags: [Genres]
      operationId: updateGenre
      summary: Update genre (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreRequest'
      responses:
        '200':
          description: Genre updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreResponse'
    delete:
      tags: [Genres]
      operationId: deleteGenre
      summary: Delete genre (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted

  # ── Stations ─────────────────────────────────────────────────────────────────
  # A Station is a themed listening room (e.g. "Rock", "Salsa").
  # It aggregates songs via its linked Genres. StationId is the room key used
  # in WebSocket presence (station:{id}) and play history.
...
  /api/v1/catalog/artists:
    get:
      tags: [Artists]
      operationId: listArtists
      summary: List artists (paginated, optional isTop filter for admin "Solo destacados")
      parameters:
        - name: isTop
          in: query
          required: false
          schema:
            type: boolean
          description: When true, returns only top/featured artists
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of artists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistPage'
    post:
      tags: [Artists]
      operationId: createArtist
      summary: Create artist (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistRequest'
      responses:
        '201':
          description: Artist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
...
  /api/v1/catalog/artists/search:
    get:
      tags: [Artists]
      operationId: searchArtists
      summary: Search artists by name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistPage'
  /api/v1/catalog/artists/{id}:
    get:
      tags: [Artists]
      operationId: getArtistById
      summary: Get artist by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Artist found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
        '404':
          description: Not found
    put:
      tags: [Artists]
      operationId: updateArtist
      summary: Update artist (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
    delete:
      tags: [Artists]
      operationId: deleteArtist
      summary: Delete artist (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted
  /api/v1/catalog/artists/{id}/albums:
    get:
      tags: [Artists]
      operationId: getArtistAlbums
      summary: Albums by artist
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPage'
  /api/v1/catalog/artists/{id}/songs:
    get:
      tags: [Artists]
      operationId: getArtistSongs
      summary: Songs by artist (top songs)
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPage'

  # ── Albums ───────────────────────────────────────────────────────────────────
  /api/v1/catalog/albums:
    get:
      tags: [Albums]
      operationId: listAlbums
      summary: List albums (paginated, optional artistId filter for admin dropdown)
      parameters:
        - name: artistId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter albums by artist (admin filter dropdown)
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPage'
    post:
      tags: [Albums]
      operationId: createAlbum
      summary: Create album (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumRequest'
      responses:
        '201':
          description: Album created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
  /api/v1/catalog/albums/search:
    get:
      tags: [Albums]
      operationId: searchAlbums
      summary: Search albums by title
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPage'
  /api/v1/catalog/albums/new-releases:
    get:
      tags: [Albums]
      operationId: getNewReleases
      summary: New releases ordered by release date
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPage'
  /api/v1/catalog/albums/top:
    get:
      tags: [Albums]
      operationId: getTopAlbums
      summary: Top albums by total streams
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumPage'
  /api/v1/catalog/albums/{id}:
    get:
      tags: [Albums]
      operationId: getAlbumById
      summary: Get album by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Album found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '404':
          description: Not found
    put:
      tags: [Albums]
      operationId: updateAlbum
      summary: Update album (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
    delete:
      tags: [Albums]
      operationId: deleteAlbum
      summary: Delete album (admin)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted
  /api/v1/catalog/albums/{id}/songs:
    get:
      tags: [Albums]
      operationId: getAlbumSongs
      summary: Songs in an album
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPage'

  # ── Songs ────────────────────────────────────────────────────────────────────
...
  /api/v1/catalog/genres/{id}/songs:
    get:
      tags: [Genres, Songs]
      operationId: getSongsByGenre
      summary: Songs tagged with a given genre (paginated)
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page of songs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongPage'

# ── Components ────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
  schemas:
    GenreRequest:
      type: object
      required: [name, gradientStart, gradientEnd]
      properties:
        name:
          type: string
          maxLength: 100
        gradientStart:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        gradientEnd:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    GenreResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        gradientStart:
          type: string
        gradientEnd:
          type: string
        songCount:
          type: integer
          description: Number of songs tagged with this genre (computed)
        createdAt:
          type: string
          format: date-time
...
    ArtistRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        photoUrl:
          type: string
        bio:
          type: string
        isTop:
          type: boolean
    ArtistResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        photoUrl:
          type: string
        bio:
          type: string
        monthlyListeners:
          type: integer
          format: int64
        followersCount:
          type: integer
          format: int64
        isTop:
          type: boolean
        createdAt:
          type: string
          format: date-time
          description: Registration date shown in admin artist detail
    ArtistPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ArtistResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
    AlbumRequest:
      type: object
      required: [title, artistId, coverUrl, releaseDate]
      properties:
        title:
          type: string
        artistId:
          type: string
          format: uuid
        coverUrl:
          type: string
        releaseDate:
          type: string
          format: date
        stationId:
          type: string
          format: uuid
          description: Optional station this album belongs to (drives "Escucha en [station]" in album detail)
    AlbumResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistResponse'
        coverUrl:
          type: string
        releaseDate:
          type: string
          format: date
        totalStreams:
          type: integer
          format: int64
        songCount:
          type: integer
          description: Number of songs in this album (computed)
        stationId:
          type: string
          format: uuid
          description: Station this album belongs to (null if not assigned)
        stationName:
          type: string
          description: Display name of the station (null if not assigned)
    AlbumPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AlbumResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
...
    SongResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          $ref: '#/components/schemas/ArtistResponse'
        album:
          $ref: '#/components/schemas/AlbumResponse'
        coverUrl:
          type: string
        audioUrl:
          type: string
        duration:
          type: integer
        lyrics:
          type: string
        playCount:
          type: integer
          format: int64
        likesCount:
          type: integer
          format: int64
        genres:
          type: array
          items:
            $ref: '#/components/schemas/GenreResponse'
    SongPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SongResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer
...